seperates config for pandas data frame as a base class
