+++ composeml/label_times/__init__.py
@@ -1,3 +1,3 @@
 # flake8:noqa
-from .data_frame import LabelTimes
 from .deserialize import read_csv, read_parquet, read_pickle
+from .object import LabelTimes
